+++ CHANGES.rst
@@ -8,6 +8,7 @@
 - Use forc ndarray flag to correctly test for fortran array contiguity [#1206]
 - Unpin ``jsonschema`` version and fix ``jsonschema`` deprecation warnings. [#1185]
 - Replace ``pkg_resources`` with ``importlib.metadata``. [#1199]
+- Unpin jsonschema, fix default validation for jsonschema 4.10+ [#1203]
 
 2.13.0 (2022-08-19)
 -------------------